requirements.txt
asgiref==3.6.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-googledrive-storage==1.6.0
django-import-export==3.1.0
django-recaptcha==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
google-api-core==2.10.2
google-api-python-client==2.81.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
protobuf==4.22.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pylibmagic==0.2.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
uritemplate==4.1.1
urllib3==1.26.15
waitress==2.1.2
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
